Running with Database
It is currently working with a local database. Used a bit more SQL to get the data I want in the correct order instead of using PHP to do it. Had to install xampp to get a local test in. I thought you could just run php in an html file and test it with a browser. I guess not. It runs in the CalendarPage HTML file now.